+++ CHANGELOG.md
@@ -1,2 +1,10 @@
-### Version 1.00 [Unreleased]
-Changelog coming soon:tm:  
+### Version 1.00 [03-03-2023]
+Added ArkInventory profile  
+Added BigWigs profile  
+Added Details profile  
+Added ElvUI layout  
+Added ElvUI aura filters  
+Added Method Raid Tools profile  
+Added OmniCD profile  
+Added Plater wago link  
+Added WeakAuras wago links